refactor: moved player back
